+++ src/StateSmith.Cli/Run/DiagramOptions.cs
@@ -1,0 +1,26 @@
+using StateSmith.Runner;
+
+namespace StateSmith.Cli.Run;
+
+public class DiagramOptions
+{
+    public TranspilerId Lang;
+    public bool NoSimGen;
+
+    public DiagramOptions()
+    {
+        
+    }
+
+    public DiagramOptions(TranspilerId lang, bool noSimGen)
+    {
+        Lang = lang;
+        NoSimGen = noSimGen;
+    }
+
+    public string Describe()
+    {
+        return $"Lang: {Lang}, NoSimGen: {NoSimGen}";
+    }
+}
+