+++ src/StateSmith.Cli/Run/RunOptions.cs
@@ -37,4 +37,11 @@ public class RunOptions
 
     [Option("no-sim-gen", HelpText = "Disables simulation .html file generation. Ignored for csx files.")]
     public bool NoSimGen { get; set; } = false;
+
+    public DiagramOptions GetDiagramOptions()
+    {
+        return new DiagramOptions(Lang, NoSimGen);
+    }
+
 }
+