+++ src/StateSmith.Cli/CHANGELOG.md
@@ -8,6 +8,13 @@ Test/interim releases are not documented here.
 
 ---
 
+## WIP
+### Added
+- `run` command now supports running diagrams with `--lang` option.
+    - https://github.com/StateSmith/StateSmith/issues/285
+
+---
+
 ## [0.8.2]
 ### Fixed
 - Fix update check showing unlisted packages when it shouldn't